+++ 1.00/PROGRAM/BS Program/SOURCE/COMMON/LM/LMM01000COMMON/DTOs/LMM01020/LMM01020DTO.cs
@@ -18,18 +18,18 @@ namespace LMM01000COMMON
         // result
         public string CCHARGES_NAME { get; set; }
         public string CCURENCY_CODE { get; set; }
-        public int NPIPE_SIZE { get; set; }
-        public int NBUY_STANDING_CHARGE { get; set; }
+        public decimal NPIPE_SIZE { get; set; }
+        public decimal NBUY_STANDING_CHARGE { get; set; }
         public string CUSAGE_RATE_MODE { get; set; }
-        public int NSTANDING_CHARGE { get; set; }
-        public int NBUY_USAGE_CHARGE_RATE { get; set; }
-        public int NUSAGE_CHARGE_RATE { get; set; }
+        public decimal NSTANDING_CHARGE { get; set; }
+        public decimal NBUY_USAGE_CHARGE_RATE { get; set; }
+        public decimal NUSAGE_CHARGE_RATE { get; set; }
         public bool LUSAGE_MIN_CHARGE { get; set; }
-        public int NUSAGE_MIN_CHARGE_AMT { get; set; }
-        public int NMAINTENANCE_FEE { get; set; }
+        public decimal NUSAGE_MIN_CHARGE_AMT { get; set; }
+        public decimal NMAINTENANCE_FEE { get; set; }
         public string CADMIN_FEE { get; set; }
-        public int NADMIN_FEE_PCT { get; set; }
-        public int NADMIN_FEE_AMT { get; set; }
+        public decimal NADMIN_FEE_PCT { get; set; }
+        public decimal NADMIN_FEE_AMT { get; set; }
         public bool LADMIN_FEE_TAX { get; set; }
         public bool LADMIN_FEE_SC { get; set; }
         public bool LADMIN_FEE_USAGE { get; set; }
